ref: Move onboarding component to correct folder
Move the create sample event component to the onboarding views folder,
since it's not a view that is reused anywhere else
