+++ src/sentry/static/sentry/app/views/onboarding/createSampleEvent.jsx
